replaced the contents of e2e-test with e2e-test-cycloud. removed e2e-test-cycloud
