+++ jobs/e2e-test.groovy
@@ -25,11 +25,11 @@ new Setup(steps
    "https://www.khanacademy.org"
 
 ).addChoiceParam(
-    "FASTLY_SERVICE",
-    """""",
-    ["PROD [VCL]", "PROD [COMPUTE]", "STAGING [COMPUTE]", "TEST [COMPUTE]"]
-)
-.addChoiceParam(
+   "FASTLY_SERVICE",
+   """""",
+   ["PROD [VCL]", "PROD [COMPUTE]", "STAGING [COMPUTE]", "TEST [COMPUTE]"]
+
+).addChoiceParam(
    "TEST_TYPE",
    """IGNORE: This is a dummy parameter that is only here to avoid breaking the
    communication with buildmaster""",
@@ -52,7 +52,7 @@ new Setup(steps
 alerts.  By default we do not send in a thread.  Generally only set by the
 buildmaster, to the 'thread_ts' or 'timestamp' value returned by the Slack
 API.""",
-    ""
+   ""
 
 ).addStringParam(
    "NUM_WORKER_MACHINES",
@@ -67,11 +67,6 @@ are, indeed, the currently active deploy.  We reserve these machines
 so the currently active deploy never has to wait for smoketest workers
 to spin up.""",
    false
-
-).addStringParam(
-   "TEST_RETRIES",
-   """How many retry attempts to use. By default is 4.""",
-   "4"
 
 ).addStringParam(
    "GIT_REVISION",
@@ -135,30 +130,54 @@ for more information.""",
 
 ).apply();
 
+// We set these to real values first thing below; but we do it within
+// the notify() so if there's an error setting them we notify on slack.
+NUM_WORKER_MACHINES = null;
+
 // Override the build name by the info that is passed in (from buildmaster).
-REVISION_DESCRIPTION = params.REVISION_DESCRIPTION ?: params.GIT_REVISION;
-E2E_URL = params.URL[-1] == '/' ? params.URL.substring(0, params.URL.length() - 1): params.URL;
+GIT_REVISION = params.GIT_REVISION;
+REVISION_DESCRIPTION = params.REVISION_DESCRIPTION ?: GIT_REVISION;
+BASE_URL = params.URL;
+E2E_URL = BASE_URL[-1] == '/' ? BASE_URL.substring(0, BASE_URL.length() - 1): BASE_URL;
 
 currentBuild.displayName = ("${currentBuild.displayName} " +
-                            "(${REVISION_DESCRIPTION})");
+   "(${REVISION_DESCRIPTION})");
 
 // We use the build name as a unique identifier for user notifications.
-BUILD_NAME = "build e2e-cypress-test #${env.BUILD_NUMBER} (${E2E_URL}: ${params.REVISION_DESCRIPTION})"
+BUILD_NAME = "${params.REVISION_DESCRIPTION} ${E2E_URL} #${env.BUILD_NUMBER}"
 
 // At this time removing @ before username.
 DEPLOYER_USER = params.DEPLOYER_USERNAME.replace("@", "")
 
 // GIT_SHA1 is the sha1 for GIT_REVISION.
 GIT_SHA1 = null;
+REPORT_DIR = "webapp/genfiles"
+REPORT_NAME = "results-combined.json"
 
 // We have a dedicated set of workers for the second smoke test.
 WORKER_TYPE = (params.USE_FIRSTINQUEUE_WORKERS
-               ? 'ka-firstinqueue-ec2' : 'ka-test-ec2');
+   ? 'ka-firstinqueue-ec2' : 'ka-test-ec2');
+
+// Used to tell whether all the test-workers raised an exception.
+public class TestFailed extends Exception {}
+
+def swallowExceptions(Closure body, Closure onException = {}) {
+   try {
+      body();
+   } catch (e) {
+      echo("Swallowing exception: ${e}");
+      onException();
+   }
+}
+
+def initializeGlobals() {
+   NUM_WORKER_MACHINES = params.NUM_WORKER_MACHINES.toInteger();
+
+   GIT_SHA1 = kaGit.resolveCommitish("git@github.com:Khan/webapp",
+      GIT_REVISION);
+}
 
 def _setupWebapp() {
-   GIT_SHA1 = kaGit.resolveCommitish("git@github.com:Khan/webapp",
-                                        params.GIT_REVISION);
-
    kaGit.safeSyncToOrigin("git@github.com:Khan/webapp", GIT_SHA1);
 
    dir("webapp/services/static") {
@@ -166,176 +185,97 @@ def _setupWebapp() {
    }
 }
 
-def runLambdaTest() {
-   // We need login creds for LambdaTest Cli
-   def lt_username = sh(script: """\
-      gcloud --project khan-academy secrets describe \
-      lambdatest_admin_account --format json | jq -r .annotations.login \
-      """, returnStdout:true).trim();
-   def lt_access_key = sh(script: """\
-      gcloud --project khan-academy secrets versions access latest \
-      --secret lambdatest_admin_account \
-      """, returnStdout:true).trim();
-
-   // Determine which environment we're running against, so we can provide a tag
-   // in the LambdaTest build.
-   def e2eEnv = E2E_URL == "https://www.khanacademy.org" ? "prod" : "preprod";
-
-   // Determine the value of the `ka-fastly-compute-environment` cookie to set
-   // when running the E2E tests.  These cookie values come from:
-   // https://khanacademy.atlassian.net/wiki/spaces/INFRA/pages/3382050914/VCL+to+Fastly+Compute+routing
-   def fastlyComputeEnvironmentCookie = "" // Default to PROD [VCL].
-   if (params.FASTLY_SERVICE == "PROD [COMPUTE]") {
-      fastlyComputeEnvironmentCookie = "NmU3ZDYxMmJjNDRkOGMwNDIzODg4ZDkyYTcxZTA4"
-   } else if (params.FASTLY_SERVICE == "STAGING [COMPUTE]") {
-      fastlyComputeEnvironmentCookie = "ODI5NTIyOGIxZTdmODUxOTEwZjY0ZDM0NjdlYjJi"
-   } else if (params.FASTLY_SERVICE == "TEST [COMPUTE]") {
-      fastlyComputeEnvironmentCookie = "ZTM3YmE2YTFjNDZkZjEwMDYxMTM3NzQyM2VlYjgw"
+// Run all the test-clients on all the worker machine, in parallel.
+def runAllTestClients() {
+   // We want to swallow any framework exceptions unless *all* the
+   // clients have raised a framework exception.  Our theory is that
+   // if one client dies unexpectedly the others can compensate, but
+   // if they all do, then there's nothing more we can do.
+   def onException = {
+      echo("Worker raised an exception");
+      WORKERS_RAISING_EXCEPTIONS++;
+      if (WORKERS_RAISING_EXCEPTIONS == NUM_WORKER_MACHINES) {
+         echo("All worker machines failed!");
+         throw new TestFailed("All worker machines failed!");
+      }
    }
 
-   def runLambdaTestArgs = ["./dev/tools/run_pkg_script.sh",
-                            "lambdatest",
-                            "--envs='FASTLY_COMPUTE_ENVIRONMENT_COOKIE=${fastlyComputeEnvironmentCookie}'",
-                            "--cy='--config baseUrl=\"${E2E_URL}\",retries=${params.TEST_RETRIES}'",
-                            "--bn='${BUILD_NAME}'",
-                            "-p=${params.NUM_WORKER_MACHINES}",
-                            "--bt='jenkins,${e2eEnv}'"
+   def jobs = [:];
+   for (i = 0; i < NUM_WORKER_MACHINES; i++) {
+      def workerId = i;  // avoid scoping problems
+      jobs["e2e-test-${workerId}"] = {
+         stage("e2e-worker-${workerId}") {
+            swallowExceptions({
+               onWorker(WORKER_TYPE, '2h') {
+                  _setupWebapp()
+                  runE2ETests(workerId)
+                  dir("${REPORT_DIR}") {
+                     stash includes: "e2e-test-results.json", name: "worker-${workerId}-reports"
+                  }
+               }
+            }, onException);
+         }
+      }
+   };
+   parallel(jobs);
+}
+
+def runE2ETests(workerId) {
+   echo("Starting e2e tests for worker ${workerId}");
+
+   // Define which environment we're running against, and setting up junit report
+   def e2eEnv = E2E_URL == "https://www.khanacademy.org" ? "prod" : "preprod";
+
+   def runE2ETestsArgs = [
+      "./dev/cypress/e2e/tools/start-cy-cloud-run.ts",
+      "--url=${E2E_URL}",
+      "--name=${BUILD_NAME}",
    ];
 
    dir('webapp/services/static') {
-      withEnv(["LT_USERNAME=${lt_username}",
-               "LT_ACCESS_KEY=${lt_access_key}"]) {
-         // NOTE: We include this Slack token in case we need to notify the
-         // FEI team when we are going to hit a `queue_timeout` error in the
-         // LambdaTest platform.
-         withCredentials([
-            string(credentialsId: "SLACK_BOT_TOKEN", variable: "SLACK_TOKEN")
-         ]) {
-            exec(runLambdaTestArgs);
+      exec(runE2ETestsArgs);
+   }
+}
+
+def unstashReports() {
+   def jsonFolders = [];
+   dir("${REPORT_DIR}") {
+      for (i = 0; i < NUM_WORKER_MACHINES; i++) {
+         exec(["rm", "-rf", "${i}"]);
+         exec(["mkdir", "-p", "${i}"]);
+         jsonFolders.add("${i}")
+         dir("./${i}") {
+            unstash "worker-${i}-reports"
+            sh("ls");
          }
       }
    }
+   return jsonFolders;
 }
 
-def runCypressCloud(){
-   build(job: 'e2e-test-cycloud',
-          parameters: [
-             string(name: 'SLACK_CHANNEL', value: "#cypress-testing"),
-             string(name: 'REVISION_DESCRIPTION', value: REVISION_DESCRIPTION),
-             string(name: 'DEPLOYER_USERNAME', value: params.DEPLOYER_USERNAME),
-             string(name: 'URL', value: E2E_URL),
-             string(name: 'NUM_WORKER_MACHINES', value: params.NUM_WORKER_MACHINES),
-             string(name: 'TEST_RETRIES', value: "1"),
-            // It takes about 5-10 minutes to run all the Cypress e2e tests when
-            // using the default of 20 workers. This build is running in parallel
-            // with Lambda tests. During this test run we don't want to disturb our 
-            // mainstream e2e pipeline, so set propagate to false.
-          ],
-          propagate: false,
-          // The pipeline will NOT wait for this job to complete to avoid
-          // blocking the main pipeline (runLambdaTest).
-          wait: false,
-          );
-}
-
-// This method filters a common 'DEPLOYER_USERNAME' into a series of comma
-// seperated slack user id's. For Example:
-// DEPLOYER_USERNAME: <@UMZGEUH09> (cc <@UN5UC0EM6>)
-// becomes: UMZGEUH09,UN5UC0EM6,
-def getUserIds(def deployUsernameBlob) {
-   // Regex to specifically grab the ids, which should start with U and be
-   // some number of capital letters and numbers. Ids can also start with
-   // W (special users), T (teams), or C (channels).
-   def pattern = /<@([UTWC][0-9A-Z]+)>/;
-   def match = (deployUsernameBlob =~ pattern);
-   if (match.results().count() == 0) {
-        return ""
+def analyzeResults(foldersList) {
+   if (currentBuild.result == 'ABORTED') {
+      // No need to report the results in the case of abort!  They will
+      // likely be more confusing than useful.
+      echo('We were aborted; no need to report results.');
+      return;
    }
 
-   def mainUser = match[0][1];
-   def otherUsers = "";
+   // report-merged-results.ts is a new file
+   kaGit.safePullInBranch("webapp/services/static/dev/cypress/e2e/tools", GIT_REVISION);
 
-   for (n in match) {
-      // We look for possible duplicates as we don't want to notify the main
-      // user twice.
-      // NOTE: This can happen when the deployer includes their name in the
-      // deploy command (e.g. `@foo` types: `sun queue foo`).
-      if (n[1] != mainUser) {
-         otherUsers += ",${n[1]}";
-      }
-   }
-
-   // Return the list of unique ids
-   return mainUser + otherUsers;
-}
-
-def analyzeResults() {
-   withTimeout('5m') {
-      if (currentBuild.result == 'ABORTED') {
-         // No need to report the results in the case of abort!  They will
-         // likely be more confusing than useful.
-         echo('We were aborted; no need to report results.');
-         return;
-      }
-
-      dir('webapp/services/static') {
-         // Get the SLACK_TOKEN from global credentials (env vars).
-         withCredentials([
-            string(credentialsId: "SLACK_BOT_TOKEN", variable: "SLACK_TOKEN")
-         ]) {
-            def notifyResultsArgs = [
-               "./dev/cypress/e2e/tools/notify-e2e-results.ts",
-               "--channel", params.SLACK_CHANNEL,
-               // The URL associated to this Jenkins build.
-               "--build-url", env.BUILD_URL,
-               // The LambdaTest build name that will be included at the
-               // beginning of the message.
-               "--label", BUILD_NAME,
-               // The URL we test against.
-               "--url", params.URL,
-               // Notify failures to DevOps and FEI
-               "--cc-on-failure", "#deploy-support-log,#fei-testing-logs"
-            ];
-
-            if (params.DEPLOYER_USERNAME) {
-               // The deployer is the person who triggered the
-               // build (this is only included in the message if
-               // the e2e tests fail).
-               notifyResultsArgs += ["--deployer", params.DEPLOYER_USERNAME];
-            }
-            if (params.SLACK_THREAD) {
-               notifyResultsArgs += ["--thread", params.SLACK_THREAD];
-            }
-
-            def userIds = getUserIds(params.DEPLOYER_USERNAME);
-
-            // Include the deployer(s) here so they can get DMs when the e2e
-            // results are ready.
-            def ccAlways = userIds ? "#cypress-logs-deploys,${userIds}" : "#cypress-logs-deploys";
-
-            if (params.SLACK_CHANNEL != "#qa-log") {
-               ccAlways += ",#qa-log";
-            }
-
-            notifyResultsArgs += ["--cc-always", ccAlways];
-
-            // notify-e2e-results returns a non-zero rc if it detects test
-            // failures. We set the job to UNSTABLE in that case (since we don't
-            // consider smoketest failures to be blocking). But we still keep on
-            // running the rest of this script!
-            catchError(buildResult: "UNSTABLE", stageResult: "UNSTABLE",
-                        message: "There were test failures!") {
-               exec(notifyResultsArgs);
-            }
-
-            // Let notify() know not to send any messages to slack, because we
-            // just did it here.
-            env.SENT_TO_SLACK = '1';
-         }
+   dir ('webapp/services/static') {
+      sh("ls ./dev/cypress/e2e/tools");
+      // report-merged-results returns a non-zero rc if it detects test
+      // failures. We set the job to UNSTABLE in that case (since we don't
+      // consider smoketest failures to be blocking). But we still keep on
+      // running the rest of this script!
+      catchError(buildResult: "UNSTABLE", stageResult: "UNSTABLE",
+         message: "There were test failures!") {
+         exec(["npx", "--yes", "tsx", "./dev/cypress/e2e/tools/report-merged-results.ts", *foldersList]);
       }
    }
 }
-
 
 // Determines if we are running the first or second smoke test.
 E2E_RUN_TYPE = (E2E_URL == "https://www.khanacademy.org" ? "second-smoke-test" : "first-smoke-test");
@@ -346,28 +286,17 @@ onWorker(WORKER_TYPE, '5h') {     // timeout
                    sender: 'Testing Turtle',
                    emoji: ':turtle:',
                    when: ['FAILURE', 'UNSTABLE']],
-           buildmaster: [sha: params.GIT_REVISION,
+           buildmaster: [sha: GIT_REVISION,
                          what: E2E_RUN_TYPE]]) {
 
-      stage("Sync webapp") {
-         _setupWebapp();
-      }
-
+      initializeGlobals();
       stage("Run e2e tests") {
-         // Note: runLambdaTest() succeeds (has an rc of 0) as long as
-         // it succeeded in running all the tests, even if some of those
-         // tests failed.  That is, this will succeed as long as there
-         // are no lambdatest framework errors; it's up to use to look
-         // for actual smoketest-code errors in analyzeResults(), below.
-         parallel([
-               "run-lambdatest": { runLambdaTest(); },
-               "test-cycloud": { runCypressCloud(); },
-            ]);
-         // runLambdaTest();
+         runAllTestClients();
       }
 
-      stage("Analyzing results") {
-         analyzeResults();
+      stage("Analyzing Results") {
+         def folders = unstashReports();
+         analyzeResults(folders);
       }
    }
 }
